update: add visualization for pcl

+++ libs/star/geometry/geometry_map/SurfelMapInitializer.cu
@@ -182,7 +182,7 @@ star::SurfelMapInitializer::~SurfelMapInitializer()
 }
 
 void star::SurfelMapInitializer::UploadDepthImage(
-    const GArrayView<unsigned> depth_image,
+    const GArrayView<unsigned short> depth_image,
     cudaStream_t stream)
 {
     // 1. Upload depth
@@ -207,7 +207,7 @@ void star::SurfelMapInitializer::UploadDepthImage(
 
 void star::SurfelMapInitializer::InitFromRGBDImage(
     const GArrayView<uchar3> color_image,
-    const GArrayView<unsigned> depth_image,
+    const GArrayView<unsigned short> depth_image,
     const float init_time,
     SurfelMap &surfel_map,
     cudaStream_t stream)
